move DO tests to one place
